handle code interpreter deltas
